Update installing.scrbl
Added method for automatic pyret download/installation. On mac, there are two things that differ from the instructions 1. the file window is not available unless a definitions window is open 2. when the link is clicked on in either safari or chrome(mac), it opens as a document in the browser instead of downloading.  There is no simple method to then download in chrome, and you have to right click on the link after opening it once to download the file on safari. 3. when the file is downloaded, it begins as a .txt file and has to be renamed in order to use in DrRacket.  This method streamlines the installation process slightly