Update_K
Update is now CPS'd and has its own ast s_update_k
Objects now have an updateWithK function, that properly handles a cpsd
update

updateWith constructs a list of applications to be made. It then creates
a continutation that will continually invoke each field and remove the
field from the list. This process will continue until the list is empty,
upon which nothing will be returned to the base k given to updateWithK.

update does not yet have the correct order of operation, it needs to be
changed from an object to a list to ensure correct ordering