CPS'ing Placeholders
Placeholders have now been cps'd in a way similar to update.
A list of guards is built up, and the continuation will continue through
them until it reaches the base guard, which will set the value and mark
that the placeholder is set.

This also fixes the remaining tests duplication bug.