External Dictionaries and Macros
Number and Strict dictionaries are now external to the runtime in
number-dict-raw.js and string-dict-raw.js respectively.

These dictionaries take advatage of the 'method' macro. It is used just
like function:

method (<self-name>, <adtl params ...>) {
    <body>
}

expands to:

makeMethod(<curried function>, <full function>)A

Each dictionary-module is given the runtime, then uses it to create the
dictionary.

The dictionary modules need to be sweetened & compiled before use, so be
sure to run make or make build to build the dictionaries after any
changes.

Tests can be run with "make test".