Fix bugs with color lists and place-image-align.
Color lists now work correctly, and exhibit the same behaviour that they
do in Racket. Part of this includes changing the definition of a natural
number from n >= 1 to n >= 0, to match Racket's definition.

place-image-align had a minor bug in the JavaScript code that caused it
to error out. It has been fixed and now works correctly.