Add target `release' for pushing builds to the CDN.
The S3 bucket `pyret-releases' will contain gzip'd builds of Pyret
(phase0 and phase1) which can be used by the editor. Note that this
depends on the command `s3', which can be downloaded here:

    http://aws.amazon.com/code/Amazon-S3/1710

The library is written in Perl, and depends on several CPAN packages,
like Net::Amazon::S3, Log::Log4perl, and Term::ShellUI. If you're
working on Brown CS machines, then these packages are already installed
for you.