Make abstract datatype description less misleading.
The description of Image, Scene and WorldConfig said that there are no
constructors available. As @blerner pointed out, this is misleading,
since they _can_ be constructed with the functions available on the
page. It is just their internal representation that cannot be
constructed. The text describing the situation is now clearer, and
refers the reader to the other available functions.