Create and use a-type-var annotation.
There is now an a-type-var annotation to mark type variable usages.
Previously, they were just converted to a-blank which meant that the
type checker couldn't see where the programmer had used declared type
variables. They are made a-type-var instead of a-name so that they can
easily be replaced with "Any" before JavaScript is emitted, like a-blank
is.