Add forall field to t-arrow and add t-var.
Instead of having a general t-forall variant, we will instead have a
field on t-arrow that introduces its type variables. We also now have a
t-var variant which takes advantage of the passed-thru a-type-var.

Type now has:
  - a _lessthan method so that types can be placed in a Set.
  - a substitute method so that types (like t-var) can be replaced
    easily

For now, the type checker will error out when two lists that should be
equal in length (such as argument types) aren't.

Finally, more builtin functions that are created by the desugarer are
handled.