Docs building as part of tests, and test-all
Add docs building to tests, and add new option, test-all

This splits make test and make test-all into separate options.  The former is
more useful to run during development for quicker responses, the latter tests
the documentation and Bootstrap as well, and is better for using on the CI
server or as a final check before pushing.