Remove foralls from t-arrow and use t-forall.
It is now possible to introduce type variables for types other than
t-forall. For example, it is now possible to write a type for something
like `empty', which has type `All(A) List<A>'.

Additionally, the type checker no longer blows up when constraints can't
be solved or go wrong. Instead a CompileError is produced saying that
the type variables can't be solved.