Add support for polymorphic data.
This commit adds support for handling polymorphic data types. Data types
can be covariant, contravariant, invariant, or bivariant (in the case of
phantom types) in their type variables. At the moment, variance is
automatically calculated for the programmer, but at some point it may be
nice to allow the programmer to specify what they want, like in OCaml:

  type +'a list = Empty | Link of 'a * 'a list;;