Preserve check fields through compilation.
The type checker will want to be able to make decisions based on check
information, so checks should be preserved through most compilation
phases. (This is preferrable to trying to recognize the pattern they
desugar to.) At some point, it may be desirable to make the desugaring
happen after the type checker has run, but this is the easiest thing to
do for now.