Hard-code commonly used modules.
Modules, default data types, and builtin types are now all wrapped in
functions inside the type-defaults.arr file. With these default values,
it is now possible to type-check a file without the "-library" flag.

Additionally, it is now possible to use check blocks, since the
list constructor that is needed can now be imported.

Some of the default types need to be cleaned up, and changed from t-top
to their actual type. This can be done over time, though.