Push through builtin imports (intermediate state, lots of logs)
This commit changes the representation of builtin modules slightly, but
support both the old and new styles.  This is really just requirejs and
JavaScript-level plumbing, nothing _new_ is being added here, just allowing
builtin modules to participate in module loading as Locators.

There are still details to work out here (ffi-helpers needs something to kick
off its module loading, for example), but I want to checkpoint in a spot where
string-dict will load in the new style before jumping to more complex things.

One shift is that, in trove/ modules (and js-defined modules in general),
Pyret and JS dependencies will no longer mix in the import line.  JS
dependencies imported with requirejs are in define as usual, but Pyret
dependencies are declared explicitly in the new "definePyretModule" form,
which is provided by runtime-util.  definePyretModule takes four arguments:

- A name for the module (the status of this name is weird, in the same way
  that the argument to util.memoModule is weird)
- A list of dependencies (described below)
- A spec of provided values (described below)
- A function of the form:

  function(runtime, namespace, pyretMod1, pyretMod2, ...) {
    ...
    return {
      // usual output stuff, including defined types, values, and checks
    }

The dependencies for now take one of two forms:

    { type: "builtin", name: String }

    { type: "dependency", protocol: String, args: RawArray<String> }

This lets them be defined in pure JS (not even depending on List<>), so that
JS modules don't need to depend on anything in Pyret.  Conversions to the
"real" datatypes used by compile-lib are specified in
compiler/locators/builtin.arr.

Right now, the provide spec is just a list of strings, and nothing uses it.
This is where type information will go, to be converted in more detail to real
types by builtin.arr's make-provides function.

The provided function is wrapped in a varargs function that closes over the
arguments and returns a function that expects runtime and namespace, and
execute's the module's body (concatenating the varargs).  This is just
convenience so that the JS module writer doesn't need to write a triple-nested
function (require modules, pyret modules, pyret runtime).

The other thing that definePyretModule does is use memoModule to avoid
evaluating the same module more than once in a runtime.  This, as always,
requires generating a decent unique name to use for caching the module's value
in a dictionary on runtime.  These names are really weird, and I'm wondering
if it would be better to somehow pass in the URI when calling the function as
a more durable unique name.  Compiled modules generate their own name with a
compile-time gensym, and it might be useful to unify those concepts.