Refactor CompileEnvironment to have separate globals and modules
The locator is responsible for knowing which globals are bound, and any
additional imports that should be added.  The compiler uses this to build a
CompileEnvironment that contains all the imported modules, with information
about what types and values are provided.

Right now the type information is minimal -- just denotes presence.  This can
be enriched and merged with type struct information down the road.