Beginning of a smarter strategy for streamlining the names and memory usage of pyret.
Because our generated JS code is in ANF, we are guaranteed that every
program-variable assignment occurs at most once.  Because Pyret is
loop-free, we are guaranteed that every function is compiled into a spinnaker
switch statement whose control flow between cases is acyclic.  This
means we can perform a live- and dead-variable analysis in linear time
without needing fixpoint analysis.

We infer the control-flow DAG between cases by recognizing assignment
statements to `$ans####`, where the particular `$ans####` name is
supplied as an argument.  (This lets us avoid confusion with
lexically-nested functions.)  We have to build this DAG after the
fact, rather than as a secondary result of `compile-fun-body` &co.,
because those compilation functions produce an incomplete case in
their results, and it's not until the full code is produced that those
incomplete cases get labels, so it's not possible to construct the DAG
eagerly.

Once we have a DAG, we traverse it a few times to produce live-in and
live-out sets and their logical complements dead-in and dead-out.  For
now, we just use the dead-out set to mark some variable names as
discardable, and back in `compile-fun-body` we use that information to
elide those names from the spilling code.  So far, this has reduced
the worst-case spilling code from 2966 vars (in trove/ast.js) down to
just 671 (in arr/compiler/type-defaults.arr.js), and reduces the
overall compiler size by 2MB, even though it adds new code.

Next steps are to actually perform "register allocation" and rename
variables that have non-interfering live ranges, and collapse the set
of names even further; and to delete dead assignments when possible.