Fix a lurking bug that manifested as a regression after 9a6ca6a2
It turns out that some .reverse() calls were missed after folding over lists
of type bindings.  Since one of them was in default-env-map-visitor, every
time this helper was used, it would flip-flop the order of type bindings in
the list.  This made it so checking unbound ids only worked after "cleaning"
the AST, which amusingly caused the required (even) number of flip-flops to
happen.  Moving the call to check-unbound broke the pattern and caused the
issue on programs like:

  type N = Number
  type N2 = N