Unify some polyglot representations to stage for bootstrapping
New field on handwritten/compiled JS modules: oldDependencies.  Used for
tracking requireJS-imported pre-instantiated modules to avoid doing module
lookup in the runtime.  This avoids loadBuiltinModule's failures.

Also, and most importantly, change the shape of compiled JS output to be an
object as created by util.definePyretModule, which contains explicit provides
and dependencies along with the module function.

TODO:
  - Build phase3/confirm clean bootstrap
  - Rewrite pyret-start/pyret.arr to use new module loading (or make new entry
    point)
  - Add back import-type, because it matters to builtin.arr
  - Confirm clean phase2/phase3 diff, but don't bootstrap yet
  - Make new bootstrapping process based on --run/--build
  - Victory?

But don't look at this commit.  It scary.