Fix builtin-modules to not error on completely empty provide specs
It should be allowed that a module specifies no provides (e.g. in
definePyretModule an empty {}).  The lack of this was causing filelib to
error.

Also, move filelib over to new style modules.

Finally, improve the error reporting in pyret-start.js for certain kinds of
JavaScript errors that are caught by Pyret's runtime, but Pyret doesn't know
how to generate error messages for.