Fixing a (long, long-standing!) performance bug in RNGLR: lengthy star items would produce quadratic numbers of reductions and SPPFNodes
because every suffix of a star item is itself a star item, and stars were translated as right-recursive rules.  The fix is to make stars be left-recursive,
and then as long as two stars are not directly next to each other (and in any case that would lead to ambiguous parsing!), reductions will take place
linearly instead of quadratically.  Thanks to @sorawee for finding a test case that demonstrated the problem clearly.