Update builtin loader to handle .arr files, and push that through
There is a new option on the CLI tool, --builtin-arr-dir, that specifies where
to find the _source_ for Pyret-implemented trove files.  This is used to find
and cache them in a new locator in locators/builtin.arr

The -no-check-mode is required because of a dependency on list in the builtin
files that is unmet if check mode is left on.  This needs to become a
per-locator decision on how to modify the compiler options to avoid check mode
in these files (the job of the old -library flag).

$ node build/phase1/main-wrapper.js \
  --builtin-js-dir src/js/troveA/ --builtin-arr-dir src/arr/troveA/ \
  -no-check-mode --build-standalone hello.arr > hello-test2.js