An initial attempt to get precise 2-way mapping from names to their usage locations
The goal is to be able to have CPO jump from the current cursor location to the beginning of the current name,
to the set of all names defined at that location, to the set of all uses of those names throughout the document,
and highlight them all.  I chose the initial character offset of the name because that's easy to obtain from CodeMirror,
whereas actual srcloc information is hard.  We need to include sets of names at each location because of datatypes,
which define Foo, is-Foo, Foo (the type) and Foo (the datatype expression) all in the same location.  I don't yet track
the definition positions of each name, though that information is easy to obtain from the resolve-name pass and can
just be passed through this new routine.

Use as `node build/phase1/main-wrapper.js src/arr/compiler/binding-info.arr <pyret-file>` for now

* Need to change s-data and s-data-expr to use Names for their data names, instead of strings.  (TODO: should t-data also take a Name?)
* Needed to make a whole bunch of locations become dummy-loc instead of inaccurate whole-program locations (No tests seem to break because of this, and I don't think an`y error reporting will get worse, but I'm not 100% certain.):
  - All the s-defined-values or defined types from a module had the entire module's location, instead of the tighter location where they were actually provided from
  - All the wrapped imports had the entire module's location, but they're phantom and should get dummy locations
  - The add-nothing-if-needed nothin had the entire module's location, but should be considered either phantom or a zero-width location at the end of the module.
* Now that s-data has a Name, needed to rename the Name that represents the entire data-expression package, to "D-Foo", for a Foo datatype, so we wouldn't get a duplicate-id error
* binding-info mimics the control flow in compile.arr, and probably should just reuse it outright, but we don't yet have a convenient way to say "compile up to name resolution and then stop"