Get some tests running again
The tests are run (for now) from tests/pyret/main2.arr

This commit skips importing all the test files that require anything more
exotic than local or built-in imports.

Some form of eval and import into the compiler would be required for the rest.

But, this does mean that we can finally have some sensible tests running again
on commit (this commit will check the travis infrastructure).