Get eval up and running again by copying handalone logic to load-lib
Several more TODOs on this, but test-include.arr proves that running
standalones from inside a Pyret process will work fine.

There's a lot of code in load-lib that needs to be sorted out;
renderErrorMessage and renderCheckResults are probably things that can become
helpers (and used by handalone), and that can be generally harmonized.