A few module loading performance improvements, and repl changes
- Only parse file locators once, and avoid evaling builtin modules until the
  contents is needed
- Split stored files into a version with just static information, and a version
  with full compiled code, to save on load time for finding dependencies

This commit also tweaks repl to take modules as an argument, to allow CPO (or
any other client of the REPL), to pass in a dictionary of starter _static_
modules.

This still needs to be expanded to allow `run-program` to take a dictionary of
starter dynamic modules, to let the repl manage its own caching.