New world-lib
Changes:

1. Resume computation when closing a big-bang window correctly
Prior this PR, when we have nested big-bangs and close the topmost big-bang window,
the computation of the previous window will not resume. Instead, the world value
of the bottommost window will be returned. Thix PR fixes this by
resuming the computation in the previous window properly.

2. Change changingWorld to a stack
Prior this PR, if the first big-bang has on-tick (which fires very frequently),
`changingWorld` could be set to true while the the second big-bang is loaded.
This shared changingWorld would cause the second big-bang to wait for the variable
to be false, but it will never be. Thus, the second big-bang will be in non-responsive state.

Setting `changingWorld` to false when resuming the computation to the previous window is also
dangerous. It's possible that the new window is populated by something like:

    on-tick(lam(_):
        call-new-big-bang()
        wait-for-an-hour()
      end)

Then, setting `changingWorld` to false could lead the callback to be called again while the old
one is still active. This could lead to an error like "Internal: run called while already running"
Thus, the best way to do this is to keep a stack of changingWorld instead.

This fixes #543

3. Add `close-when-stop` per a discussion with @joe
A new WorldConfig which could be used to close the big-bang window when `stop-when` returns true.

4. Enforce modal dialog correctly
When a modal dialog appears, it should grey out everything else, including other dialog.
This is the correct functionality of JQuery UI modal dialog.
Prior this PR manual changes to z-index in Pyret make all dialogs accessible. This PR fixes
this issue.

Note that one nice consequence is that we do not need to worry whether
the current computation / window should be halted / closed if users close the previous
window because the correct modal dialog will force users to interact with only the topmost window.

5. Prevent computation resuming until the dialog is closed
It doesn't make sense to resume the computation while the dialog is open
(even when the world has stoped) because the computation could then trigger a new big-bang
which will prevent viewing the old window which is the only reason we do not close the window
right away when the big-bang stops.

6. Clean big-bang box when closed
This fixes #483.

This PR should be merged along with <CPO PR>

With this PR, the following program should be able to run properly:
https://code.pyret.org/editor#share=0Bxr4FfLa3goOREVhM1hRMUJtWW8