Minimal changes for an experiment in moving list operations to JS
This isn't the most efficient way to do this by a long shot (each(), in
particular, doesn't need to allocate a new array), but it consistently makes a
clean `make phaseB` go from about 310-320 seconds down to 270-280 seconds for
me.  It also doesn't cover all the methods used in the compiler.