Adding tracing of all toplevel values to the compilation process.
Running a program can opt-in to recording all the values, simply by setting the `currentURL` parameter on the runtime to the URL of the program being run, and setting the `onRecordTrace` parameter to true.
Or, running a program can register its own callback for each value, by setting the currentURL parameter as above, and setting the `onRecordTrace` parameter to a callback function.  This function will run on the Pyret stack of the program-being-run, so may need to safeCall or pauseStack/run in order to do anything that might be stack-heavy.