Refactoring unbound-id and bad-assignment checking -- cleaning up semantics and speeding up compile times.
It used to make sense that ast-util could check for bad assignments or
unbound ids, because we didn't yet have any place that tracked scope
particularly well.  But resolve-scope is designed to do exactly that,
and it has its own notions of bindings that are in scope.  Moreover,
because name resolution turns s-names into s-atoms, s-globals or
s-type-globals, we're guaranteed that each binding site is of a unique
Name.  Therefore, the binding information being tracked by
resolve-scope should suffice to decide unbound-ids and
bad-assignments, without needing ast-util to reconstruct all that
information.  Some details of what information is tracked are slightly
different (ast-util tried to deduce a bit of aliasing information with
some heuristics), but they don't affect the results.

One consequence of relying on the binding information from
resolve-scope is that any new bindings generated by desugaring
(especially by the underscore-functions) must be added to that store
of bindings, or else spurious unbound-id errors will happen.

Another, potentially-unneeded but eminently plausible change made is
to move unbound-id and bad-assignment checking earlier, before
type-checking.  Why should the type-checker have to deal with unknown
names?

I merged the two checkers into a single visitor, saving an additional
pass over the AST.  This saves ~10% of compile time on my machine.