Start building up better data structures for static cross-module info
Most directly, this commit fixes #288.

It also gives a place to start hanging more static information off of module
exports.  The information is all "there" but needs to be attached to particular
bindings in different phases.

This should eventually help with statically checking cases (#366) and attaching
cross-module locations to bindings, or associating bindings with import lines
(#678).  The `ValueExport` type also captures the flatness analysis information
that @puppyofkosh is working on.

I'm not sure that the `ValueExport` datatype is in its final form, and how many
variants it will eventually have.  But it's a starting point that makes
variables work, and demonstrated how to get more than just type information
propagated around across modules.  In particular, it maybe should be unified
with the the `ScopeBinding` type, since they reflect similar information, just
cross- vs intra-module.

This commit also uses the "defined-values" field instead of the object-shaped
"values" in "provide-plus-types".  This is because, with variables exported,
it's impossible to represent an exported variable in a Pyret object, since a
Pyret object's field could only hold the value of the variable.  Here, the cell
itself is exported.

I think this is actually a surprisingly expressive feature for a module system
to offer, though the alternative is to rework how the REPL works entirely to
not use module imports and exports as the semantics of binding between entries.