Intentionally committing a slightly broken version for illustration
This version makes re-exporting a shadowed global import not work (hence some
failing tests).

It also highlights a deeper issue – there's no way, using ScopeBinding, to
talk about a binding that is both _global_ and a _variable_.  The reason this
matters is that it affects what values are considered to be "defined" within
the module.  So the current hacky check for the value being present in the
globals.values dictionary doesn't cover the case where a module re-defines a
value that came from a global.  If that check is omitted, then global mutable
variables across the REPL are always attempted to be provided, which is
obviously wrong, and causes other errors.

I think ScopeBinding needs to be refactored to have a "where" and a "what", so
the "where" can be a module, local, or global, and the "what" can be a
variable, let-binding, or recursive binding.