More progress on this, but still not totally fixed
This has shifted the issue around – this commit _disables_ a new, useful
error that would be good to catch.

Right now, if you write:

x :: Number.y = 10

This passes well-formedness and scope resolution, because the scope resolver
doesn't take into account which type names are module names and which type
names are type names.

This is easy to fix in the whole-module situation, but at the REPL, it's
trickier.  If the main program writes:

import string-dict as SD

Then we need to know what SD is in the interactions window.  The REPL can't
just add the same import statement at the top of every entry, because this
leads to issues with shadowing.

So "SD" needs to be added to the REPL environment as more than just any old
type or value – it should be a _module_ type or value.  This is important also
for making sure that different kinds of cases compilation work well.

As a result, I think there may be one more kind of export a module can provide
(aside from values, types, and new datatypes), which is _modules_ that are
re-exported.  This is useful also for modules that "package up" other modules,
so you might write:

include data-analysis # which re-exports math, statistics, and probability

math.d-dx(lam(x): x * x end)
statistics.mean([list: 1, 2, 3])
d :: probability.Distribution = probability.normal(10, 2)

And the whole first-order system would know that math, statistics, and
probability are all modules that may contain their own types, datatypes,
values, etc.