Very basic server for handling compile requests
Example usage:

$ make src/arr/compiler/server.jarr
$ node src/arr/compiler/server.jarr
< logs stuff about server >

(elsewhere)

$ node src/arr/compiler/client.js --program <something>.arr --require-config src/scripts/standalone-configA.json

For now, both need to be run from the pyret-lang directory, and only the server
process logs anything useful.  Errors aren't pretty printed.

But, some compile errors are effectively instantaneous, which is a nice
improvement.  This should be straightforward to iterate on, with INFO messages
back to the client for progress, and pretty-printed errors, etc, rendered and
passed across.