Avoid using requirejs for building standalones
Why?  Because requirejs has a custom parser, and uses it to parse all the
JavaScript that flows through it.  That includes HUGE files like the generated
Pyret parser.  (Note that Pyret already manually managed files generated from
Pyret, so those aren't considered).  As a result, this saves about 5-8
_seconds_ on my machine during the "Cleaning up and generating standalone..."
step of the compiler.  This should make recompiling the compiler on small edits
much more pleasant.

The main change is twofold:

- All pure JS modules that are "nativeRequires" need to have a unique global
  name.  Since we control these modules, this is easy to do.  That's what all
  the changes in the top "define" calls are, along with the changes to the
  paths in the parser generator.
- We need to have a configuration file that _lists_ all the paths to these
  files explicitly, since we no longer rely on requirejs to find them.  That's
  what's going on in standalone-configABC

This is currently compiled with a flag coming from the config file --
"use-raw-files".  Depending on the future of this, the case in make-standalone
that checks this could be omitted entirely, or we could keep the requirejs
option available.