Avoid cross-namespace shadowing issues, add tests
Because it's not clear what the right design decision is on shadowing across
types, modules, and values, this commit “turns off” that checking.

We can turn it back on or make it an option, but this keeps things
backwards-compatible for now, so that a breaking change isn't tied in with the
way this is implemented.