A very rough sketch of using Node's repl library
To run, from the root of the REPO, run:

$ make src/arr/compiler/cli-repl.jarr
$ node src/arr/compiler/cli-repl.jarr
> x = 10
PBase { dict: {}, brands: { brandCount: 0 } }
> x
10

Obviously, the
result.result.dict.v.val.result.result.dict.answer is
ridiculous. The levels of unwrapping here are several:

- result.result: Accessing the value within the result of the
  run from run-interaction. It's a SuccessResult, which means
  that nothing went wrong internally running this (a
  FailureResult would mean some module-level, internal
  misconfiguration)
- ....dict.v: The result was an Either, so we access the v
  field (assuming it was a right() representing successful
  run). If it were a left(), it would mean there was a compile
  error
- ....val: The answer in the either is a POpaque value
  containing some goodies inside. These are defined in
  load-lib, which defines what a ModuleResult is.
- ....result.result: This is actually the result of running
  the module as a value. It contains more than just the final
  answer value, and includes defined names and types, and any
  test results
- ....dict.answer: One of the fields on the module answer is
  called answer, and this accesses that value

There isn't a great wrapper that takes all of the myriad
possibilities here (which are compile error, runtime error,
and successful run); CPO defines its own multi-level
extraction of this information to render the different kinds
of errors. It would be great to write this in a way that lets
us program against this more sensibly.

Aside from that, this is straightforward use of APIs that aren't
that well documented. The incantations in cli-repl.arr are
lifted from test-repl.arr and lightly adapted.

There's a bit of interesting runtime management going on in
cli-repl.js. Note that cli-repl.arr calls start(), which calls
pauseStack immediately and sets things up so only on exit is the
start() calling-context re-started. This is because the REPL
library is about to run off and create a bunch of event handlers
and who knows what, the original Pyret program that started it
may or may not be done. So, much like in a reactor, we pause the
stack and then restart it in the 'exit' event. Note that this
means that the return value from start() in JS is irrelevant; it
is never used. Only the restarter.resume() callback matters to
the Pyret program.

Note also that it's worth looking into how SIGINT gets handled
by the "repl" library, because we may be able to make it mean
"stop the current run", not just "stop the whole interaction"