New entrypoint for caching idea for worklists
There are three steps to module loading:

1 locate all the modules
2 compile all the modules
3 instantiate all the modules

We have good caching strategies for 2 & 3, but not for 1. This adds a
new entrypoint for building the worklist that allows the worklist to
ignore traversing modules selected by the caller. The caller is
responsible for knowing that they are already available in the static
cache, and will be provided to compile-module later.