REPL fixed by using exports and imports of module names
Finally! This completes the third namespace for modules, and uses them in a
principled way within repl.arr to pass them around that works with the
type-checker.

This also has the seeds of a more complete treatment of ProvideBlocks, which
can be developed along with a more complete treatment of include blocks.

There are some new opportunities for good error messages, and for consolidating
runtime lookups of types and ids that came from modules, as well.