Rework s-include and s-import-complete to use s-import-from
This re-works the internal desugaring of import syntax to use s-include-from,
and adds the start of s-import-from to scope resolution.

A high-level point is that this is moving away from the model of include/import
xxx from being an import plus let statements, and instead adding the
appropriate names to the environment directly, then relying on the
representation of those identifiers to compile them correctly. This is
intentional and has an eye towards making the module system model a first-order
alias analysis of provided names, so we can always tell when two imports are
actually the same underlying provided value or type. This helps with various
compilation efforts, errors, and keeping things canonical so the type checker
has good notions of type equality.

Many TODOs remain, including processing renamings, checking shadowing in all
cases, and processing the meaning of importing entire data definitions.