Make flatness respect module bindings
This includes both matching on a-id-modref in the a-let case for when a user
wrote

    import lists as L
    is-link = L.is-link

and on a-id-modref in an a-app:

    lists.is-link(foo)

In addition, this changes the way datatypes are put into provides to make them
use their datatype plain name instead of an atom. This matches how global.js
and string-dict.js have (perhaps happened to) work for a long time, as well as
how the type checker generates provides.

If types start shadowing one another we may need to revisit this. However,
within provides it doesn't make sense for a module to export an alias and a
datatype with the same name, so from outside the module this is a moot point;
it's just inside the module that we would need to figure out which one "wins".

This also changes cmdline to export only names it defines; once `provide ...
as` lands we can switch it back to using the new syntax.